Refactor Telegram gateway into focused modules
